--- runhttp.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-"""Gate HTTP API server runner."""
-
-#from gateserver import db, http_api
-#import config
-
-#if __name__ == '__main__':
-    #db.connect(config.db_url)
-    #http_api.serve(config)
-print('Not implemented yet.')
